Move new hub example tests to the models package
**What**
- Move these to the models package so that integrations with code
  coverage in VS Code is improved. It allows us to run the package tests
  and then view the missed lines from within the editor directly
