+++ pkg/generators/models/testdata/cases/allof1/api.yaml
