+++ pkg/generators/models/testdata/cases/allof1/expected/model_foo.go
