+++ pkg/generators/models/testdata/cases/allof2/api.yaml
