+++ pkg/generators/models/testdata/cases/allof2/expected/model_foo.go
