+++ pkg/generators/models/testdata/cases/allof3/api.yaml
