+++ pkg/generators/models/testdata/cases/allof3/expected/model_foo.go
