+++ pkg/generators/models/testdata/cases/allof3/expected/model_sub.go
