+++ pkg/generators/models/testdata/cases/allof4/api.yaml
