+++ pkg/generators/models/testdata/cases/allof4/expected/model_foo.go
