+++ pkg/generators/models/testdata/cases/embedded_type/api.yaml
