+++ pkg/generators/models/testdata/cases/embedded_type/expected/model_sub.go
