+++ pkg/generators/models/testdata/cases/embedded_type/expected/model_top.go
