+++ pkg/generators/models/testdata/cases/nullable_arrays/api.yaml
