+++ pkg/generators/models/testdata/cases/nullable_arrays/expected/model_foo.go
