+++ pkg/generators/models/testdata/cases/nullable_properties/api.yaml
