+++ pkg/generators/models/testdata/cases/nullable_properties/expected/model_foo.go
