+++ pkg/generators/models/testdata/cases/nullable_untyped_object/api.yaml
