+++ pkg/generators/models/testdata/cases/nullable_untyped_object/expected/model_foo.go
