+++ pkg/generators/models/testdata/cases/object_with_additional_properties/api.yaml
