+++ pkg/generators/models/testdata/cases/oneof/api.yaml
