+++ pkg/generators/models/testdata/cases/oneof/expected/model_foo.go
