+++ pkg/generators/models/testdata/cases/required_properties/api.yaml
