+++ pkg/generators/models/testdata/cases/required_properties/expected/model_foo.go
