+++ pkg/generators/models/testdata/cases/simple_model/api.yaml
