+++ pkg/generators/models/testdata/cases/simple_model/expected/model_foo.go
