+++ pkg/generators/models/testdata/cases/typed_arrays/api.yaml
