+++ pkg/generators/models/testdata/cases/typed_arrays/expected/model_foo.go
