+++ pkg/generators/models/testdata/cases/untyped_object/api.yaml
