+++ pkg/generators/models/testdata/cases/untyped_object/expected/model_foo.go
