+++ pkg/generators/models/testdata/model_sub_type.go
@@ -7,7 +7,7 @@ package testpkg
 
 // SubType is an object.
 type SubType struct {
-	// Foo
+	// Foo:
 	Foo string `json:"foo,omitempty"`
 }
